Add files via upload
The above A1 to A9 are the AIML technical projects developed, coded and executed by me. A  more elaborative work is the customer churn AI capstone project wherein the python functions have been coded to detect the churn rate of customers. This project can be extended on real time industry pipelines and can be used as unit testing.